remove the test that don't make sense anymore.
At the time I wrote this test, Moose-2.0402 was still on CPAN, so we
were sure that's the distribution we want. Now they are both deleted,
and we are not sure which one is the one we want. Because PAUSE will
index the distribution with the same version numbers in them, there's
no way we can figure out the older one is the accurate one.